measure the tile error based on a vector pointing horizontally away
this doesn't work well:
- currently the AABBs have huge top/down margins. after clipping the
nearest point might be in a horizontal plane with the camera. that
results in a very small error for that tile, and the wrong zoom level is
loaded.
- even with better AABBs, the problem would remain, as that vector might
indicate a small error while in reality we are looking at the slope of a
mountain (90 degrees)

however, the number of tiles loaded was just too large with the
previous, more conservative bound (horizontal x-axis).

more information is necessary to do it good. at the very least, we would
need much improved AABBs. one option would be to then to calculate the
error based on 4 triangles with a base on the bottom side of the AABB,
and the peak in the middle at the top side.

Another option would be to do it like the others, using rendered data.
maybe that isn't that slow after all.