initial version of render thread notifier. several things unclear:
- first and foremost: why are tiles still appearing after stopping touch
and no touch? it seems like the render loop keeps running for as long as
is needed for the tiles to be loaded. this doesn't make sense. maybe a
qt internal workaround (network thread waking up the render thread?).
- how to perform a clean startup. initialisation is a mess atm.
- what's the order of qml init? it seems that the render thread is
started up before the qml fires QQmlApplicationEngine::objectCreated ->
logical, it is started during setup.
- what is the qobject::connect in the main thread for? looks like a
workaround for sth (was in place in an example or sth). but the root
window is in place after load finishes. maybe that depends on sth?